test: break up util test files
