--- __tests__/support-item-utils.test.ts
@@ -1,24 +1,0 @@
-import {
-	getSupportCategories,
-	getRegistrationGroups,
-	getSupportItemDefs,
-} from "@/lib/support-item-utils";
-
-describe("support item utils", () => {
-	it("should get support categories", () => {
-		const supportCategories = getSupportCategories();
-		expect(supportCategories.length).toEqual(15);
-	});
-
-	it("should get registration groups", () => {
-		const registrationGroups = getRegistrationGroups();
-		expect(registrationGroups.length).toEqual(35);
-	});
-
-	it("should get support items", () => {
-		const supportItems = getSupportItemDefs(
-			"Group And Centre Based Activities"
-		);
-		expect(supportItems.length).toEqual(104);
-	});
-});
